Fixing the array to value conversion
IN our test case there is a time the pram is a key => array. Something like this

```
Array
(
    [context] => test
    [return_type] => ids
    [mappings] => Array
        (
            [0] => Array
                (
                    [attributes] => Array
                        (
                            [0] => Array
                                (
                                    [name] => role
                                    [value] => value
                                )

                        )

                )

        )

)
```

Then the pulgin crash.

This change will test if it is an array then skip. The content is purely a reportportal content. I am not sure if it is good to convert them in full in terms of retaining information. Because if we allow that and if the param is too long then the description might be too large.

Open for suggestion. For now this fixes out test case.