Excepting existing launch
In our use case we already have one launch (details is that we create one
launch per pipeline ID) and would like to put all suites test under that launch
as root item. This allows us to have a view by pipeline ID.

This change together with this PR
https://github.com/Mikalai-Kabzar/client-php/pull/1 would allow us to get this
done.

The reason why we just create the launch in codecept PHP is that the launch has
nothing to do with the PHP codeception test, it is related to the pipeline
rather.

Note that this will fail if the above PR has not been merged. One can use my repo instead in the branch `go1custom` in the mean time.

https://github.com/sunshine69/client-php